% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_plot_alice.R
\name{obtn_plot_alice}
\alias{obtn_plot_alice}
\title{Title}
\usage{
obtn_plot_alice(
  obtn_year,
  county_to_plot,
  plot_width = 3.1,
  plot_height = 0.3574
)
}
\arguments{
\item{plot_height}{}
}
\value{

}
\description{
Title
}
