% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_plot_choropleth_map.R
\name{obtn_plot_choropleth_map}
\alias{obtn_plot_choropleth_map}
\title{Make Choropleth Map}
\usage{
obtn_plot_choropleth_map(
  obtn_year,
  measure_to_plot,
  plot_width = 4.3684,
  plot_height = 3.25
)
}
\arguments{
\item{obtn_year}{}
}
\value{

}
\description{
Make Choropleth Map
}
