% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_plot_median_income.R
\name{obtn_plot_median_income}
\alias{obtn_plot_median_income}
\title{Make a single median income plot}
\usage{
obtn_plot_median_income(
  obtn_year,
  county_to_plot,
  plot_width = 2.43,
  plot_height = 0.61
)
}
\arguments{
\item{obtn_year}{}
}
\value{

}
\description{
Make a single median income plot
}
\examples{



}
