% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_plot_population_pyramid.R
\name{obtn_plot_population_pyramid}
\alias{obtn_plot_population_pyramid}
\title{Title}
\usage{
obtn_plot_population_pyramid(
  obtn_year,
  county_to_plot,
  plot_width = 3.27,
  plot_height = 4.0752
)
}
\arguments{
\item{plot_height}{}
}
\value{

}
\description{
Title
}
