% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_plot_race_ethnicity_bar_chart.R
\name{obtn_plot_race_ethnicity_bar_chart}
\alias{obtn_plot_race_ethnicity_bar_chart}
\title{Title}
\usage{
obtn_plot_race_ethnicity_bar_chart(
  obtn_year,
  county_to_plot,
  plot_width = 3.1,
  plot_height = 2.1
)
}
\arguments{
\item{county_to_plot}{}
}
\value{

}
\description{
Title
}
