% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_plot_top_employment_industries.R
\name{obtn_plot_top_employment_industries}
\alias{obtn_plot_top_employment_industries}
\title{Make Map of Top Employment Industries in Oregon Counties}
\usage{
obtn_plot_top_employment_industries(
  obtn_year,
  industry_to_plot,
  plot_width = 2,
  plot_height = 1.5
)
}
\arguments{
\item{obtn_year}{Year to plot}

\item{industry_to_plot}{Industry to plot}

\item{plot_width}{Plot width}

\item{plot_height}{Plot height}
}
\value{

}
\description{
Make Map of Top Employment Industries in Oregon Counties
}
