% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtn_save_plot.R
\name{obtn_save_plot}
\alias{obtn_save_plot}
\title{Title}
\usage{
obtn_save_plot(obtn_year, measure, geography, plot_width, plot_height)
}
\arguments{
\item{plot_height}{}
}
\value{

}
\description{
Title
}
